mark oog_static_memory dummy& add test
