+++ zkevm-circuits/src/evm_circuit/execution/pop.rs
@@ -105,4 +105,27 @@ mod test {
     fn pop_gadget_rand() {
         test_ok(rand_word());
     }
+
+    fn test_stack_underflow(value: Word) {
+        let bytecode = bytecode! {
+            PUSH32(value)
+            POP
+            POP
+            STOP
+        };
+
+        assert_eq!(
+            run_test_circuits(
+                TestContext::<2, 1>::simple_ctx_with_bytecode(bytecode).unwrap(),
+                None
+            ),
+            Ok(())
+        );
+    }
+
+    #[test]
+    fn pop_gadget_underflow() {
+        test_stack_underflow(Word::from(0x030201));
+        test_stack_underflow(Word::from(0xab));
+    }
 }
